+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Events from external at update.
+
 ## [1.0.0-alpha.0] - 2022-12-07
 
 ### Added